Moves new tests to new file
