+++ pageserve/test_routes.py
@@ -1,0 +1,29 @@
+"""Copyright 2019 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+
+import unittest
+from unittest.mock import patch
+import app
+
+
+class TestRoutes(unittest.TestCase):
+    def setUp(self):
+        """Set up test client."""
+        self.client = app.APP.test_client()
+
+    def test_index(self):
+        """Check that index page is rendered correctly."""
+        self.client.get('/v1/')
+        self.assert_template_used('index.html')
